Moved some functionality out of main.
