+++ src/main.rs
@@ -1,8 +1,5 @@
-use sqlx::PgPool;
-use std::net::TcpListener;
 use zero2prod::configuration::get_configuration;
-use zero2prod::mail::EmailClient;
-use zero2prod::startup::run;
+use zero2prod::startup::build;
 use zero2prod::telemetry::{get_subscriber, init_subscriber};
 
 #[tokio::main]
@@ -14,26 +11,7 @@ async fn main() -> std::io::Result<()> {
     // Configuration
     let configuration = get_configuration().expect("Failed to get configuration");
 
-    // SQL Database setup
-    let db_connection = PgPool::connect_lazy_with(configuration.database.with_db());
-
-    // TCP Listener setup for App
-    let address = format!("{}:{}", configuration.app.host, configuration.app.port);
-    let listener = TcpListener::bind(address)?;
-
-    // Email Client Setup
-    let sender = configuration
-        .email_client
-        .sender()
-        .expect("Failed to parse sender email");
-    let timeout = configuration.email_client.timeout();
-    let email_client = EmailClient::new(
-        sender,
-        configuration.email_client.base_url,
-        configuration.email_client.auth_token,
-        timeout,
-    );
-
-    // FIRE!
-    run(listener, db_connection, email_client)?.await
+    let server = build(configuration)?;
+    server.await?;
+    Ok(())
 }
